# low collision + fast mixer, more optimization