@gunnarmorling 2024-01-19
Wow, this comes out on top for the standard data set!

```
Benchmark 1: timeout -v 300 ./calculate_average_abeobk.sh 2>&1
  Time (mean ± σ):      2.487 s ±  0.047 s    [User: 16.786 s, System: 0.703 s]
  Range (min … max):    2.445 s …  2.609 s    10 runs

Summary
  abeobk: trimmed mean 2.47728426946, raw times 2.48665502846,2.48285004746,2.4671632004600004,2.47460331246,2.47546025346,2.4452292854600004,2.51112508146,2.60882662046,2.4701938394600003,2.4502233924600003

Leaderboard

| # | Result (m:s.ms) | Implementation     | JDK | Submitter     | Notes     |
|---|-----------------|--------------------|-----|---------------|-----------|
|   | 00:02.477 | [link](https://github.com/gunnarmorling/1brc/blob/main/src/main/java/dev/morling/onebrc/CalculateAverage_abeobk.java)| 21.0.1-graal | [Van Phu DO](https://github.com/abeobk) | GraalVM native binary |
```

CC @royvanrijn, @thomaswue, @artsiomkorzun