@abeobk 2024-01-20
Thx, I realized that we can return immediately when max is found, therefore checking for min is then not necessary.